Pull most of the server-side harness for client testing into `testing-server`.
