+++ local-modules/testing-server/package.json
@@ -6,6 +6,7 @@
     "chai": "^4.1.0",
     "chai-as-promised": "^7.1.1",
     "mocha": "^3.3.0",
+    "puppeteer": "^0.11.0",
 
     "env-server": "local",
     "util-common": "local"
