+++ server/package.json
@@ -17,7 +17,6 @@
 
   "dependencies": {
     "minimist": "^1.2.0",
-    "puppeteer": "^0.11.0",
     "source-map-support": "^0.5.0",
 
     "app-setup": "local",
